集成到单一文件
