+++ package.json
@@ -23,5 +23,8 @@
   "devDependencies": {
     "dtslint": "Microsoft/dtslint#production",
     "types-publisher": "Microsoft/types-publisher#production"
+  },
+  "dependencies": {
+    "jslint": "^0.10.3"
   }
 }
